# Add error to Decoder.PeekKind and add Decoder.More