Add an error return variable to PeekKind.
This avoids accidentally ignoring an error condition during peeking.

Add a Decoder.More method to compensate for loss of ergonomics. The More method can be used to quickly iterate through all elements or members of a JSON array or object.
This mirrors the existing jsonv1.Decoder.More method. Switch existing usages of PeekKind to use More.

Performance:

    name                                           old time/op    new time/op    delta
    Testdata/CanadaGeometry/Unmarshal/Concrete     1.20ms ± 1%    1.20ms ± 1%    ~     (p=0.310 n=5+5)
    Testdata/CanadaGeometry/Unmarshal/Interface    1.64ms ± 0%    1.66ms ± 1%  +1.18%  (p=0.008 n=5+5)
    Testdata/CitmCatalog/Unmarshal/Concrete        2.04ms ± 1%    2.13ms ± 0%  +4.37%  (p=0.008 n=5+5)
    Testdata/CitmCatalog/Unmarshal/Interface       3.88ms ± 1%    4.02ms ± 1%  +3.63%  (p=0.008 n=5+5)
    Testdata/GolangSource/Unmarshal/Concrete       5.51ms ± 0%    5.70ms ± 0%  +3.54%  (p=0.008 n=5+5)
    Testdata/GolangSource/Unmarshal/Interface      9.09ms ± 1%    9.60ms ± 1%  +5.59%  (p=0.008 n=5+5)
    Testdata/SyntheaFhir/Unmarshal/Concrete        3.60ms ± 1%    3.56ms ± 0%  -1.34%  (p=0.008 n=5+5)
    Testdata/SyntheaFhir/Unmarshal/Interface       5.33ms ± 1%    5.40ms ± 1%  +1.26%  (p=0.016 n=5+5)
    Testdata/TwitterStatus/Unmarshal/Concrete      1.16ms ± 1%    1.13ms ± 0%  -2.94%  (p=0.008 n=5+5)
    Testdata/TwitterStatus/Unmarshal/Interface     2.02ms ± 1%    2.02ms ± 1%    ~     (p=0.421 n=5+5)