@dsnet 2025-01-24
I'm not sold that this is a good idea, but I wanted to see what it would take to make `PeekKind` return an error.