@dsnet 2025-01-25
I posted my thoughts here: https://github.com/golang/go/discussions/63397#discussioncomment-11948805

The more I look at it, the more unconvinced I am that `PeekKind` needs to return an error.