@mvdan 2025-01-27
I originally asked about PeekKind returning an error by looking at two particular use cases that would be improved by it, but from what you've written about other reasonable use cases, and the performance, overall I agree that I lean against adding an error return to PeekKind.

Like you however, I would improve the docs a bit so that they are clearer. I'd also like to see documented what PeekKind does when a non-EOF error is encountered, such as invalid syntax.